backend/pyproject.toml
[tool.poetry]
name = "plsom-backend"
version = "0.0.1"
description = "Backend For PLSOM LMS"
authors = ["Dave Mcsavvy <davemcsavvii@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.4"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-backblaze-b2 = "^7.0.0"
sentry-sdk = {extras = ["django"], version = "^2.29.1"}
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"
python-decouple = "^3.8"
pillow = "^11.2.1"
celery = "^5.5.3"
redis = "^6.2.0"
psycopg2-binary = "^2.9.10"
django-unfold = "^0.59.0"
drf-spectacular = "^0.28.0"
dj-database-url = "^2.3.0"
whitenoise = "^6.9.0"
mypy = "^1.16.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.31.0"
mypy = "^1.15.0"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    ".vscode",
    ".ruff_cache",
    ".mypy_cache",
    ".pytest_cache",
    "*.ipynb"
]

[tool.mypy]
ignore_missing_imports = true